requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
openai-agents>=0.0.7
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
